Split up consensus module
